+++ src/main/java/com/example/hyper/dtos/requests/ArtistRequestDTO.java
@@ -1,11 +1,11 @@
-package com.example.hyper.dtos;
+package com.example.hyper.dtos.requests;
 
 import com.example.hyper.entities.BaseEntity;
 import lombok.Data;
 import javax.validation.constraints.NotEmpty;
 
 @Data
-public class ArtistDTO extends BaseEntity {
+public class ArtistRequestDTO extends BaseEntity {
 
     private Long id;
