+++ src/main/java/com/example/hyper/dtos/requests/UserRequestDTO.java
@@ -1,4 +1,4 @@
-package com.example.hyper.dtos;
+package com.example.hyper.dtos.requests;
 
 import com.example.hyper.entities.BaseEntity;
 import lombok.Data;
@@ -6,7 +6,7 @@ import lombok.Data;
 import javax.validation.constraints.NotEmpty;
 
 @Data
-public class UserDTO extends BaseEntity {
+public class UserRequestDTO extends BaseEntity {
 
     private Long id;
