+++ src/main/java/com/example/hyper/dtos/responses/ArtistPageResponseDTO.java
@@ -1,6 +1,6 @@
 package com.example.hyper.dtos.responses;
 
-import com.example.hyper.dtos.ArtistDTO;
+import com.example.hyper.dtos.requests.ArtistRequestDTO;
 import com.example.hyper.dtos.MusicDTO;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
@@ -26,5 +26,5 @@ public class ArtistPageResponseDTO {
     private Integer number;
 
     @JsonProperty("artistDataList")
-    private List<ArtistDTO> content;
+    private List<ArtistRequestDTO> content;
 }
