+++ src/main/java/com/example/hyper/services/ArtistService.java
@@ -1,21 +1,18 @@
 package com.example.hyper.services;
 
-import com.example.hyper.dtos.ArtistDTO;
-import com.example.hyper.dtos.MusicDTO;
+import com.example.hyper.dtos.requests.ArtistRequestDTO;
 import com.example.hyper.dtos.responses.ArtistPageResponseDTO;
 import com.example.hyper.dtos.responses.ArtistResponseDTO;
-import com.example.hyper.dtos.responses.MusicPageResponseDTO;
-import com.example.hyper.dtos.responses.MusicResponseDTO;
 import org.springframework.data.domain.Pageable;
 
 import java.util.List;
 
 public interface ArtistService {
 
-    public ArtistResponseDTO save(ArtistDTO artist);
+    public ArtistResponseDTO save(ArtistRequestDTO artist);
 
     public ArtistPageResponseDTO find(List<String> names, Pageable pageable);
 
-    public ArtistResponseDTO update(Long id, ArtistDTO artist);
+    public ArtistResponseDTO update(Long id, ArtistRequestDTO artist);
     public void delete(Long id);
 }
