+++ src/main/java/com/example/hyper/services/UserService.java
@@ -1,6 +1,6 @@
 package com.example.hyper.services;
 
-import com.example.hyper.dtos.UserDTO;
+import com.example.hyper.dtos.requests.UserRequestDTO;
 import com.example.hyper.dtos.responses.UserPageResponseDTO;
 import com.example.hyper.dtos.responses.UserResponseDTO;
 import org.springframework.data.domain.Pageable;
@@ -9,11 +9,11 @@ import java.util.List;
 
 public interface UserService {
 
-    public UserResponseDTO save(UserDTO user);
+    public UserResponseDTO save(UserRequestDTO user);
 
     public UserPageResponseDTO find(List<String> names, Pageable pageable);
 
-    public UserResponseDTO update(Long id, UserDTO user);
+    public UserResponseDTO update(Long id, UserRequestDTO user);
 
     public void delete(Long id);
 }
